refactor: HistoryList 내부 유틸함수 분리
