+++ src/HistoryList.tsx
@@ -1,25 +1,7 @@
 import { FC } from "react";
 import { css } from "@emotion/react";
 
-const convertDateFormat = (num: number): string => {
-  const dateFormat = new Date(num);
-  const year = dateFormat.getFullYear() % 100;
-  const month = (dateFormat.getMonth() + 1).toLocaleString("en-US", {
-    minimumIntegerDigits: 2,
-    useGrouping: false,
-  });
-  const date = dateFormat
-    .getDate()
-    .toLocaleString("en-US", { minimumIntegerDigits: 2, useGrouping: false });
-  const hours = dateFormat
-    .getHours()
-    .toLocaleString("en-US", { minimumIntegerDigits: 2, useGrouping: false });
-  const minutes = dateFormat
-    .getMinutes()
-    .toLocaleString("en-US", { minimumIntegerDigits: 2, useGrouping: false });
-
-  return `${year}-${month}-${date} ${hours}:${minutes}`;
-};
+import { convertToDateFormat } from "./utils";
 
 interface IHistoryListProps {
   histories: chrome.history.HistoryItem[];
@@ -64,7 +46,7 @@ const HistoryList: FC<IHistoryListProps> = ({ histories }) => {
             `}
           >
             {history.lastVisitTime
-              ? convertDateFormat(history.lastVisitTime)
+              ? convertToDateFormat(history.lastVisitTime)
               : "정보없음"}
           </span>
         </li>
